refactor(prover-client): collapse orchestrator into checkpoint + top-tree schedulers
Replace the old ProvingOrchestrator + EpochProvingState + EpochProvingContext +
TopTreeProvingScheduler stack with a checkpoint-driven design.

- `CheckpointSubTreeOrchestrator` extends `ProvingScheduler` directly, taking
  `ChonkCache + EpochNumber` instead of the EpochProvingContext indirection.
- `TopTreeOrchestrator` extends `ProvingScheduler` (collapses the former
  TopTreeProvingScheduler).
- `CheckpointProvingState` takes three discrete deps: `epochNumber`, `isAlive`,
  `onReject` — no CheckpointParent interface, no parentEpoch.
- `ChonkCache`: tx-hash-keyed cache of chonk-verifier proofs, released when the
  owning epoch expires.

Deletes the old orchestrator unit tests; new behaviour is covered per-component
in the new prover-node tests landing in subsequent commits.
