+++ yarn-project/prover-client/src/mocks/test_context.ts
@@ -5,13 +5,13 @@ import { BlockNumber, CheckpointNumber } from '@aztec/foundation/branded-types';
 import { padArrayEnd, times, timesAsync } from '@aztec/foundation/collection';
 import { Fr } from '@aztec/foundation/curves/bn254';
 import type { Logger } from '@aztec/foundation/log';
+import { SerialQueue } from '@aztec/foundation/queue';
 import type { FieldsOf } from '@aztec/foundation/types';
 import { getVKTreeRoot } from '@aztec/noir-protocol-circuits-types/vk-tree';
 import { ProtocolContractsList } from '@aztec/protocol-contracts';
 import { computeFeePayerBalanceLeafSlot } from '@aztec/protocol-contracts/fee-juice';
 import { PublicDataWrite } from '@aztec/stdlib/avm';
 import { AztecAddress } from '@aztec/stdlib/aztec-address';
-import { EthAddress } from '@aztec/stdlib/block';
 import type { Checkpoint } from '@aztec/stdlib/checkpoint';
 import type { MerkleTreeWriteOperations, ServerCircuitProver } from '@aztec/stdlib/interfaces/server';
 import type { CheckpointConstantData } from '@aztec/stdlib/rollup';
@@ -36,11 +36,20 @@ import {
   getTreeSnapshot,
   insertSideEffects,
 } from '../orchestrator/block-building-helpers.js';
-import type { BlockProvingState } from '../orchestrator/block-proving-state.js';
-import { ProvingOrchestrator } from '../orchestrator/index.js';
 import { BrokerCircuitProverFacade } from '../proving_broker/broker_prover_facade.js';
 import { TestBroker } from '../test/mock_prover.js';
 import { getEnvironmentConfig, getSimulator, makeCheckpointConstants, makeGlobals } from './fixtures.js';
+
+/**
+ * Builds a started `SerialQueue` for use as an orchestrator's deferred-job queue in
+ * tests. Production wires a single shared queue via `ProverClient`; tests that construct
+ * orchestrators directly use one of these per orchestrator (or share one).
+ */
+export function makeTestDeferredJobQueue(concurrency = 10): SerialQueue {
+  const queue = new SerialQueue();
+  queue.start(concurrency);
+  return queue;
+}
 
 export class TestContext {
   private headers: Map<number, BlockHeader> = new Map();
@@ -57,17 +66,12 @@ export class TestContext {
     public prover: ServerCircuitProver,
     public broker: TestBroker,
     public brokerProverFacade: BrokerCircuitProverFacade,
-    public orchestrator: TestProvingOrchestrator,
     private feePayer: AztecAddress,
     initialFeePayerBalance: Fr,
     private directoriesToCleanup: string[],
     private logger: Logger,
   ) {
     this.feePayerBalance = initialFeePayerBalance;
-  }
-
-  public get epochProver() {
-    return this.orchestrator;
   }
 
   static async new(
@@ -118,22 +122,11 @@ export class TestContext {
 
     const broker = new TestBroker(proverCount, localProver);
     const facade = new BrokerCircuitProverFacade(broker);
-    const orchestrator = new TestProvingOrchestrator(ws, facade, EthAddress.ZERO, false, 10);
 
     await broker.start();
     facade.start();
 
-    return new this(
-      ws,
-      localProver,
-      broker,
-      facade,
-      orchestrator,
-      feePayer,
-      initialFeePayerBalance,
-      directoriesToCleanup,
-      logger,
-    );
+    return new this(ws, localProver, broker, facade, feePayer, initialFeePayerBalance, directoriesToCleanup, logger);
   }
 
   public getFork() {
@@ -350,18 +343,5 @@ export class TestContext {
     }
 
     return endStateReference;
-  }
-}
-
-class TestProvingOrchestrator extends ProvingOrchestrator {
-  public isVerifyBuiltBlockAgainstSyncedStateEnabled = false;
-
-  // Disable this check by default, since it requires seeding world state with the block being built
-  // This is only enabled in some tests with multiple blocks that populate the pending chain via makePendingBlock
-  protected override verifyBuiltBlockAgainstSyncedState(provingState: BlockProvingState): Promise<void> {
-    if (this.isVerifyBuiltBlockAgainstSyncedStateEnabled) {
-      return super.verifyBuiltBlockAgainstSyncedState(provingState);
-    }
-    return Promise.resolve();
   }
 }
