--- yarn-project/prover-client/src/orchestrator/orchestrator.ts
@@ -1,1111 +1,0 @@
-import { BatchedBlob, FinalBlobBatchingChallenges, SpongeBlob } from '@aztec/blob-lib/types';
-import {
-  L1_TO_L2_MSG_SUBTREE_HEIGHT,
-  L1_TO_L2_MSG_SUBTREE_ROOT_SIBLING_PATH_LENGTH,
-  NESTED_RECURSIVE_ROLLUP_HONK_PROOF_LENGTH,
-  NUM_BASE_PARITY_PER_ROOT_PARITY,
-} from '@aztec/constants';
-import { BlockNumber, EpochNumber } from '@aztec/foundation/branded-types';
-import { Fr } from '@aztec/foundation/curves/bn254';
-import type { LoggerBindings } from '@aztec/foundation/log';
-import { promiseWithResolvers } from '@aztec/foundation/promise';
-import { assertLength } from '@aztec/foundation/serialize';
-import { pushTestData } from '@aztec/foundation/testing';
-import { elapsed } from '@aztec/foundation/timer';
-import type { TreeNodeLocation } from '@aztec/foundation/trees';
-import { EthAddress } from '@aztec/stdlib/block';
-import type {
-  EpochProver,
-  ForkMerkleTreeOperations,
-  MerkleTreeWriteOperations,
-  PublicInputsAndRecursiveProof,
-  ReadonlyWorldStateAccess,
-  ServerCircuitProver,
-} from '@aztec/stdlib/interfaces/server';
-import { appendL1ToL2MessagesToTree } from '@aztec/stdlib/messaging';
-import type { Proof } from '@aztec/stdlib/proofs';
-import {
-  type BaseRollupHints,
-  BlockRootEmptyTxFirstRollupPrivateInputs,
-  BlockRootFirstRollupPrivateInputs,
-  BlockRootSingleTxFirstRollupPrivateInputs,
-  BlockRootSingleTxRollupPrivateInputs,
-  CheckpointConstantData,
-  CheckpointRootSingleBlockRollupPrivateInputs,
-  PrivateTxBaseRollupPrivateInputs,
-  PublicChonkVerifierPrivateInputs,
-  PublicChonkVerifierPublicInputs,
-  RootRollupPublicInputs,
-} from '@aztec/stdlib/rollup';
-import type { CircuitName } from '@aztec/stdlib/stats';
-import { type AppendOnlyTreeSnapshot, MerkleTreeId } from '@aztec/stdlib/trees';
-import type { BlockHeader, ProcessedTx, Tx } from '@aztec/stdlib/tx';
-import type { UInt64 } from '@aztec/stdlib/types';
-import {
-  Attributes,
-  type TelemetryClient,
-  type Tracer,
-  getTelemetryClient,
-  trackSpan,
-  wrapCallbackInSpan,
-} from '@aztec/telemetry-client';
-
-import { inspect } from 'util';
-
-import {
-  buildHeaderFromCircuitOutputs,
-  getLastSiblingPath,
-  getPublicChonkVerifierPrivateInputsFromTx,
-  getRootTreeSiblingPath,
-  getSubtreeSiblingPath,
-  getTreeSnapshot,
-  insertSideEffectsAndBuildBaseRollupHints,
-  validatePartialState,
-  validateTx,
-} from './block-building-helpers.js';
-import type { BlockProvingState } from './block-proving-state.js';
-import type { CheckpointProvingState } from './checkpoint-proving-state.js';
-import { EpochProvingState, type ProvingResult, type TreeSnapshots } from './epoch-proving-state.js';
-import { ProvingOrchestratorMetrics } from './orchestrator_metrics.js';
-import { TopTreeProvingScheduler } from './top-tree-proving-scheduler.js';
-import { TxProvingState } from './tx-proving-state.js';
-
-/**
- * Implements an event driven proving scheduler to build the recursive proof tree. The idea being:
- * 1. Transactions are provided to the scheduler post simulation.
- * 2. Tree insertions are performed as required to generate transaction specific proofs
- * 3. Those transaction specific proofs are generated in the necessary order accounting for dependencies
- * 4. Once a transaction is proven, it will be incorporated into a merge proof
- * 5. Merge proofs are produced at each level of the tree until the root proof is produced
- *
- * The proving implementation is determined by the provided prover. This could be for example a local prover or a remote prover pool.
- */
-
-/**
- * The orchestrator, managing the flow of recursive proving operations required to build the rollup proof tree.
- */
-export class ProvingOrchestrator extends TopTreeProvingScheduler implements EpochProver {
-  protected provingState: EpochProvingState | undefined = undefined;
-
-  protected provingPromise: Promise<ProvingResult> | undefined = undefined;
-  private metrics: ProvingOrchestratorMetrics;
-
-  private dbs: Map<BlockNumber, MerkleTreeWriteOperations> = new Map();
-
-  constructor(
-    private dbProvider: ReadonlyWorldStateAccess & ForkMerkleTreeOperations,
-    prover: ServerCircuitProver,
-    private readonly proverId: EthAddress,
-    private readonly cancelJobsOnStop: boolean = false,
-    enqueueConcurrency: number,
-    telemetryClient: TelemetryClient = getTelemetryClient(),
-    bindings?: LoggerBindings,
-  ) {
-    super(prover, enqueueConcurrency, 'prover-client:orchestrator', bindings);
-    this.metrics = new ProvingOrchestratorMetrics(telemetryClient, 'ProvingOrchestrator');
-  }
-
-  get tracer(): Tracer {
-    return this.metrics.tracer;
-  }
-
-  public getProverId(): EthAddress {
-    return this.proverId;
-  }
-
-  public getNumActiveForks() {
-    return this.dbs.size;
-  }
-
-  protected override cancelInternal(): void {
-    this.cancel();
-  }
-
-  protected override wrapCircuitCall<T>(
-    circuitName: string,
-    fn: (signal: AbortSignal) => Promise<T>,
-  ): (signal: AbortSignal) => Promise<T> {
-    return wrapCallbackInSpan(
-      this.tracer,
-      `ProvingOrchestrator.prover.${circuitName}`,
-      { [Attributes.PROTOCOL_CIRCUIT_NAME]: circuitName as CircuitName },
-      fn,
-    );
-  }
-
-  protected override onRootRollupComplete(state: EpochProvingState) {
-    state.resolve({ status: 'success' });
-  }
-
-  public startNewEpoch(
-    epochNumber: EpochNumber,
-    totalNumCheckpoints: number,
-    finalBlobBatchingChallenges: FinalBlobBatchingChallenges,
-  ) {
-    if (this.provingState?.verifyState()) {
-      throw new Error(
-        `Cannot start epoch ${epochNumber} when epoch ${this.provingState.epochNumber} is still being processed.`,
-      );
-    }
-
-    const { promise: _promise, resolve, reject } = promiseWithResolvers<ProvingResult>();
-    const promise = _promise.catch((reason): ProvingResult => ({ status: 'failure', reason }));
-    this.logger.info(`Starting epoch ${epochNumber} with ${totalNumCheckpoints} checkpoints.`);
-    this.provingState = new EpochProvingState(
-      epochNumber,
-      totalNumCheckpoints,
-      finalBlobBatchingChallenges,
-      provingState => this.checkAndEnqueueCheckpointRootRollup(provingState),
-      resolve,
-      reject,
-    );
-    this.provingPromise = promise;
-  }
-
-  /**
-   * Starts a new checkpoint.
-   * @param checkpointIndex - The index of the checkpoint in the epoch.
-   * @param constants - The constants for this checkpoint.
-   * @param l1ToL2Messages - The set of L1 to L2 messages to be inserted at the beginning of this checkpoint.
-   * @param totalNumBlocks - The total number of blocks expected in the checkpoint (must be at least one).
-   * @param headerOfLastBlockInPreviousCheckpoint - The header of the last block in the previous checkpoint.
-   */
-  public async startNewCheckpoint(
-    checkpointIndex: number,
-    constants: CheckpointConstantData,
-    l1ToL2Messages: Fr[],
-    totalNumBlocks: number,
-    headerOfLastBlockInPreviousCheckpoint: BlockHeader,
-  ) {
-    if (!this.provingState) {
-      throw new Error('Empty epoch proving state. Call startNewEpoch before starting a checkpoint.');
-    }
-
-    if (!this.provingState.isAcceptingCheckpoints()) {
-      throw new Error(`Epoch not accepting further checkpoints.`);
-    }
-
-    // Fork world state at the end of the immediately previous block.
-    const lastBlockNumber = headerOfLastBlockInPreviousCheckpoint.globalVariables.blockNumber;
-    const db = await this.dbProvider.fork(lastBlockNumber);
-
-    const firstBlockNumber = BlockNumber(lastBlockNumber + 1);
-    this.dbs.set(firstBlockNumber, db);
-
-    // Get archive sibling path before any block in this checkpoint lands.
-    const lastArchiveSiblingPath = await getLastSiblingPath(MerkleTreeId.ARCHIVE, db);
-
-    // Insert all the l1 to l2 messages into the db. And get the states before and after the insertion.
-    const {
-      lastL1ToL2MessageTreeSnapshot,
-      lastL1ToL2MessageSubtreeRootSiblingPath,
-      newL1ToL2MessageTreeSnapshot,
-      newL1ToL2MessageSubtreeRootSiblingPath,
-    } = await this.updateL1ToL2MessageTree(l1ToL2Messages, db);
-
-    this.provingState.startNewCheckpoint(
-      checkpointIndex,
-      constants,
-      totalNumBlocks,
-      headerOfLastBlockInPreviousCheckpoint,
-      lastArchiveSiblingPath,
-      l1ToL2Messages,
-      lastL1ToL2MessageTreeSnapshot,
-      lastL1ToL2MessageSubtreeRootSiblingPath,
-      newL1ToL2MessageTreeSnapshot,
-      newL1ToL2MessageSubtreeRootSiblingPath,
-    );
-  }
-
-  /**
-   * Starts off a new block
-   * @param blockNumber - The block number
-   * @param timestamp - The timestamp of the block. This is only required for constructing the private inputs for the
-   * block that doesn't have any txs.
-   * @param totalNumTxs - The total number of txs in the block
-   */
-  @trackSpan('ProvingOrchestrator.startNewBlock', blockNumber => ({
-    [Attributes.BLOCK_NUMBER]: blockNumber,
-  }))
-  public async startNewBlock(blockNumber: BlockNumber, timestamp: UInt64, totalNumTxs: number) {
-    if (!this.provingState) {
-      throw new Error('Empty epoch proving state. Call startNewEpoch before starting a block.');
-    }
-
-    const checkpointProvingState = this.provingState.getCheckpointProvingStateByBlockNumber(blockNumber);
-    if (!checkpointProvingState) {
-      throw new Error(`Checkpoint not started. Call startNewCheckpoint first.`);
-    }
-
-    if (!checkpointProvingState.isAcceptingBlocks()) {
-      throw new Error(`Checkpoint not accepting further blocks.`);
-    }
-
-    const constants = checkpointProvingState.constants;
-    this.logger.info(`Starting block ${blockNumber} for slot ${constants.slotNumber}.`);
-
-    // Fork the db only when it's not already set. The db for the first block is set in `startNewCheckpoint`.
-    if (!this.dbs.has(blockNumber)) {
-      // Fork world state at the end of the immediately previous block
-      const db = await this.dbProvider.fork(BlockNumber(blockNumber - 1));
-      this.dbs.set(blockNumber, db);
-    }
-    const db = this.getDbForBlock(blockNumber);
-
-    // Get archive snapshot and sibling path before any txs in this block lands.
-    const lastArchiveTreeSnapshot = await getTreeSnapshot(MerkleTreeId.ARCHIVE, db);
-    const lastArchiveSiblingPath = await getRootTreeSiblingPath(MerkleTreeId.ARCHIVE, db);
-
-    const blockProvingState = checkpointProvingState.startNewBlock(
-      blockNumber,
-      timestamp,
-      totalNumTxs,
-      lastArchiveTreeSnapshot,
-      lastArchiveSiblingPath,
-    );
-
-    // Enqueue base parity circuits for the first block in the checkpoint.
-    if (blockProvingState.index === 0) {
-      for (let i = 0; i < NUM_BASE_PARITY_PER_ROOT_PARITY; i++) {
-        this.enqueueBaseParityCircuit(checkpointProvingState, blockProvingState, i);
-      }
-    }
-
-    // Because `addTxs` won't be called for a block without txs, and that's where the sponge blob state is computed.
-    // We need to set its end sponge blob here, which will become the start sponge blob for the next block.
-    if (totalNumTxs === 0) {
-      const endState = await db.getStateReference();
-      blockProvingState.setEndState(endState);
-
-      const endSpongeBlob = blockProvingState.getStartSpongeBlob().clone();
-      const blockEndBlobFields = blockProvingState.getBlockEndBlobFields();
-      await endSpongeBlob.absorb(blockEndBlobFields);
-      blockProvingState.setEndSpongeBlob(endSpongeBlob);
-
-      // Try to accumulate the out hashes and blobs as far as we can:
-      await this.provingState.accumulateCheckpointOutHashes();
-      await this.provingState.setBlobAccumulators();
-    }
-  }
-
-  /**
-   * The interface to add simulated transactions to the scheduler. This can only be called once per block.
-   * @param txs - The transactions to be proven
-   */
-  @trackSpan('ProvingOrchestrator.addTxs', txs => ({
-    [Attributes.BLOCK_TXS_COUNT]: txs.length,
-  }))
-  public async addTxs(txs: ProcessedTx[]): Promise<void> {
-    if (!this.provingState) {
-      throw new Error(`Empty epoch proving state. Call startNewEpoch before adding txs.`);
-    }
-
-    if (!txs.length) {
-      // To avoid an ugly throw below. If we require an empty block, we can just call setBlockCompleted
-      // on a block with no txs. We cannot do that here because we cannot find the blockNumber without any txs.
-      this.logger.warn(`Provided no txs to orchestrator addTxs.`);
-      return;
-    }
-
-    const blockNumber = BlockNumber(txs[0].globalVariables.blockNumber);
-    const provingState = this.provingState.getBlockProvingStateByBlockNumber(blockNumber!);
-    if (!provingState) {
-      throw new Error(`Proving state for block ${blockNumber} not found. Call startNewBlock first.`);
-    }
-
-    if (provingState.totalNumTxs !== txs.length) {
-      throw new Error(
-        `Block ${blockNumber} should be filled with ${provingState.totalNumTxs} txs. Received ${txs.length} txs.`,
-      );
-    }
-
-    if (!provingState.isAcceptingTxs()) {
-      throw new Error(`Block ${blockNumber} has been initialized with transactions.`);
-    }
-
-    this.logger.info(`Adding ${txs.length} transactions to block ${blockNumber}`);
-
-    const db = this.getDbForBlock(blockNumber);
-    const lastArchive = provingState.lastArchiveTreeSnapshot;
-    const newL1ToL2MessageTreeSnapshot = provingState.newL1ToL2MessageTreeSnapshot;
-    const spongeBlobState = provingState.getStartSpongeBlob().clone();
-
-    for (const tx of txs) {
-      try {
-        if (!provingState.verifyState()) {
-          throw new Error(`Invalid proving state when adding a tx`);
-        }
-
-        validateTx(tx);
-
-        this.logger.debug(`Received transaction: ${tx.hash}`);
-
-        const startSpongeBlob = spongeBlobState.clone();
-        const [hints, treeSnapshots] = await this.prepareBaseRollupInputs(
-          tx,
-          lastArchive,
-          newL1ToL2MessageTreeSnapshot,
-          startSpongeBlob,
-          db,
-        );
-
-        if (!provingState.verifyState()) {
-          throw new Error(`Unable to add transaction, preparing base inputs failed`);
-        }
-
-        await spongeBlobState.absorb(tx.txEffect.toBlobFields());
-
-        const txProvingState = new TxProvingState(tx, hints, treeSnapshots, this.proverId.toField());
-        const txIndex = provingState.addNewTx(txProvingState);
-        if (txProvingState.requireAvmProof) {
-          this.getOrEnqueueChonkVerifier(provingState, txIndex);
-          this.logger.debug(`Enqueueing public VM for tx ${txIndex}`);
-          this.enqueueVM(provingState, txIndex);
-        } else {
-          this.logger.debug(`Enqueueing base rollup for private-only tx ${txIndex}`);
-          this.enqueueBaseRollup(provingState, txIndex);
-        }
-      } catch (err: any) {
-        throw new Error(`Error adding transaction ${tx.hash.toString()} to block ${blockNumber}: ${err.message}`, {
-          cause: err,
-        });
-      }
-    }
-
-    const endState = await db.getStateReference();
-    provingState.setEndState(endState);
-
-    const blockEndBlobFields = provingState.getBlockEndBlobFields();
-    await spongeBlobState.absorb(blockEndBlobFields);
-
-    provingState.setEndSpongeBlob(spongeBlobState);
-
-    // Txs have been added to the block. Now try to accumulate the out hashes and blobs as far as we can:
-    await this.provingState.accumulateCheckpointOutHashes();
-    await this.provingState.setBlobAccumulators();
-  }
-
-  /**
-   * Kickstarts chonk verifier circuits for the specified txs. These will be used during epoch proving.
-   * Note that if the chonk verifier circuits are not started this way, they will be started nontheless after processing.
-   */
-  @trackSpan('ProvingOrchestrator.startChonkVerifierCircuits')
-  public startChonkVerifierCircuits(txs: Tx[]) {
-    if (!this.provingState?.verifyState()) {
-      throw new Error(`Empty epoch proving state. call startNewEpoch before starting chonk verifier circuits.`);
-    }
-    const publicTxs = txs.filter(tx => tx.data.forPublic);
-    for (const tx of publicTxs) {
-      const txHash = tx.getTxHash().toString();
-      const privateInputs = getPublicChonkVerifierPrivateInputsFromTx(tx, this.proverId.toField());
-      const tubeProof =
-        promiseWithResolvers<
-          PublicInputsAndRecursiveProof<
-            PublicChonkVerifierPublicInputs,
-            typeof NESTED_RECURSIVE_ROLLUP_HONK_PROOF_LENGTH
-          >
-        >();
-      this.logger.debug(`Starting chonk verifier circuit for tx ${txHash}`);
-      this.doEnqueueChonkVerifier(txHash, privateInputs, proof => {
-        tubeProof.resolve(proof);
-      });
-      this.provingState.cachedChonkVerifierProofs.set(txHash, tubeProof.promise);
-    }
-    return Promise.resolve();
-  }
-
-  /**
-   * Marks the block as completed.
-   * Computes the block header and updates the archive tree.
-   */
-  @trackSpan('ProvingOrchestrator.setBlockCompleted', (blockNumber: BlockNumber) => ({
-    [Attributes.BLOCK_NUMBER]: blockNumber,
-  }))
-  public async setBlockCompleted(blockNumber: BlockNumber, expectedHeader?: BlockHeader): Promise<BlockHeader> {
-    const provingState = this.provingState?.getBlockProvingStateByBlockNumber(blockNumber);
-    if (!provingState) {
-      throw new Error(`Block proving state for ${blockNumber} not found`);
-    }
-
-    // Abort with specific error for the block if there's one.
-    const error = provingState.getError();
-    if (error) {
-      throw new Error(`Block proving failed: ${error}`);
-    }
-
-    // Abort if the proving state is not valid due to errors occurred elsewhere.
-    if (!provingState.verifyState()) {
-      throw new Error(`Invalid proving state when completing block ${blockNumber}.`);
-    }
-
-    if (provingState.isAcceptingTxs()) {
-      throw new Error(
-        `Block ${blockNumber} is still accepting txs. Call setBlockCompleted after all txs have been added.`,
-      );
-    }
-
-    // Given we've applied every change from this block, now assemble the block header:
-    this.logger.verbose(`Block ${blockNumber} completed. Assembling header.`);
-    const header = await provingState.buildBlockHeader();
-
-    if (expectedHeader && !header.equals(expectedHeader)) {
-      this.logger.error(`Block header mismatch: header=${header} expectedHeader=${expectedHeader}`);
-      throw new Error('Block header mismatch');
-    }
-
-    // Get db for this block and remove from map — no other code should use it after this point.
-    const db = this.getDbForBlock(provingState.blockNumber);
-    this.dbs.delete(provingState.blockNumber);
-
-    // Update the archive tree, capture the snapshot, and close the fork deterministically.
-    try {
-      this.logger.verbose(
-        `Updating archive tree with block ${provingState.blockNumber} header ${(await header.hash()).toString()}`,
-      );
-      await db.updateArchive(header);
-      provingState.setBuiltArchive(await getTreeSnapshot(MerkleTreeId.ARCHIVE, db));
-    } finally {
-      await db.close();
-    }
-
-    await this.verifyBuiltBlockAgainstSyncedState(provingState);
-
-    return header;
-  }
-
-  // Flagged as protected to disable in certain unit tests
-  protected async verifyBuiltBlockAgainstSyncedState(provingState: BlockProvingState) {
-    const builtBlockHeader = provingState.getBuiltBlockHeader();
-    if (!builtBlockHeader) {
-      this.logger.debug('Block header not built yet, skipping header check.');
-      return;
-    }
-
-    const output = provingState.getBlockRootRollupOutput();
-    if (!output) {
-      this.logger.debug('Block root rollup proof not built yet, skipping header check.');
-      return;
-    }
-
-    const newArchive = provingState.getBuiltArchive();
-    if (!newArchive) {
-      this.logger.debug('Archive snapshot not yet captured, skipping header check.');
-      return;
-    }
-
-    const header = await buildHeaderFromCircuitOutputs(output);
-
-    if (!(await header.hash()).equals(await builtBlockHeader.hash())) {
-      this.logger.error(`Block header mismatch.\nCircuit: ${inspect(header)}\nComputed: ${inspect(builtBlockHeader)}`);
-      provingState.reject(`Block header hash mismatch.`);
-      return;
-    }
-
-    const blockNumber = provingState.blockNumber;
-    const syncedArchive = await getTreeSnapshot(MerkleTreeId.ARCHIVE, this.dbProvider.getSnapshot(blockNumber));
-    if (!syncedArchive.equals(newArchive)) {
-      this.logger.error(
-        `Archive tree mismatch for block ${blockNumber}: world state synced to ${inspect(
-          syncedArchive,
-        )} but built ${inspect(newArchive)}`,
-      );
-      provingState.reject(`Archive tree mismatch.`);
-      return;
-    }
-
-    const circuitArchive = output.newArchive;
-    if (!newArchive.equals(circuitArchive)) {
-      this.logger.error(`New archive mismatch.\nCircuit: ${output.newArchive}\nComputed: ${newArchive}`);
-      provingState.reject(`New archive mismatch.`);
-      return;
-    }
-  }
-
-  /**
-   * Cancel any further proving.
-   * If cancelJobsOnStop is true, aborts all pending jobs with the broker (which marks them as 'Aborted').
-   * If cancelJobsOnStop is false (default), jobs remain in the broker queue and can be reused on restart/reorg.
-   */
-  public cancel() {
-    this.resetSchedulerState(this.cancelJobsOnStop);
-
-    this.provingState?.cancel();
-
-    for (const [blockNumber, db] of this.dbs.entries()) {
-      void db.close().catch(err => this.logger.error(`Error closing db for block ${blockNumber}`, err));
-    }
-    this.dbs.clear();
-  }
-
-  private getDbForBlock(blockNumber: BlockNumber): MerkleTreeWriteOperations {
-    const db = this.dbs.get(blockNumber);
-    if (!db) {
-      throw new Error(`World state fork for block ${blockNumber} not found.`);
-    }
-    return db;
-  }
-
-  /**
-   * Returns the proof for the current epoch.
-   */
-  public async finalizeEpoch(): Promise<{
-    publicInputs: RootRollupPublicInputs;
-    proof: Proof;
-    batchedBlobInputs: BatchedBlob;
-  }> {
-    if (!this.provingState || !this.provingPromise) {
-      throw new Error(`Invalid proving state, an epoch must be proven before it can be finalized`);
-    }
-
-    const result = await this.provingPromise!;
-    if (result.status === 'failure') {
-      throw new Error(`Epoch proving failed: ${result.reason}`);
-    }
-
-    await this.provingState.finalizeBatchedBlob();
-
-    const epochProofResult = this.provingState.getEpochProofResult();
-
-    pushTestData('epochProofResult', {
-      proof: epochProofResult.proof.toString(),
-      publicInputs: epochProofResult.publicInputs.toString(),
-    });
-
-    return epochProofResult;
-  }
-
-  private async updateL1ToL2MessageTree(l1ToL2Messages: Fr[], db: MerkleTreeWriteOperations) {
-    const lastL1ToL2MessageTreeSnapshot = await getTreeSnapshot(MerkleTreeId.L1_TO_L2_MESSAGE_TREE, db);
-    const lastL1ToL2MessageSubtreeRootSiblingPath = assertLength(
-      await getSubtreeSiblingPath(MerkleTreeId.L1_TO_L2_MESSAGE_TREE, L1_TO_L2_MSG_SUBTREE_HEIGHT, db),
-      L1_TO_L2_MSG_SUBTREE_ROOT_SIBLING_PATH_LENGTH,
-    );
-
-    // Update the local trees to include the new l1 to l2 messages
-    await appendL1ToL2MessagesToTree(db, l1ToL2Messages);
-
-    const newL1ToL2MessageTreeSnapshot = await getTreeSnapshot(MerkleTreeId.L1_TO_L2_MESSAGE_TREE, db);
-    const newL1ToL2MessageSubtreeRootSiblingPath = assertLength(
-      await getSubtreeSiblingPath(MerkleTreeId.L1_TO_L2_MESSAGE_TREE, L1_TO_L2_MSG_SUBTREE_HEIGHT, db),
-      L1_TO_L2_MSG_SUBTREE_ROOT_SIBLING_PATH_LENGTH,
-    );
-
-    return {
-      lastL1ToL2MessageTreeSnapshot,
-      lastL1ToL2MessageSubtreeRootSiblingPath,
-      newL1ToL2MessageTreeSnapshot,
-      newL1ToL2MessageSubtreeRootSiblingPath,
-    };
-  }
-
-  // Updates the merkle trees for a transaction. The first enqueued job for a transaction
-  @trackSpan('ProvingOrchestrator.prepareBaseRollupInputs', tx => ({
-    [Attributes.TX_HASH]: tx.hash.toString(),
-  }))
-  private async prepareBaseRollupInputs(
-    tx: ProcessedTx,
-    lastArchive: AppendOnlyTreeSnapshot,
-    newL1ToL2MessageTreeSnapshot: AppendOnlyTreeSnapshot,
-    startSpongeBlob: SpongeBlob,
-    db: MerkleTreeWriteOperations,
-  ): Promise<[BaseRollupHints, TreeSnapshots]> {
-    // We build the base rollup inputs using a mock proof and verification key.
-    // These will be overwritten later once we have proven the chonk verifier circuit and any public kernels
-    const [ms, hints] = await elapsed(
-      insertSideEffectsAndBuildBaseRollupHints(
-        tx,
-        lastArchive,
-        newL1ToL2MessageTreeSnapshot,
-        startSpongeBlob,
-        this.proverId.toField(),
-        db,
-      ),
-    );
-
-    this.metrics.recordBaseRollupInputs(ms);
-
-    const promises = [MerkleTreeId.NOTE_HASH_TREE, MerkleTreeId.NULLIFIER_TREE, MerkleTreeId.PUBLIC_DATA_TREE].map(
-      async (id: MerkleTreeId) => {
-        return { key: id, value: await getTreeSnapshot(id, db) };
-      },
-    );
-    const treeSnapshots: TreeSnapshots = new Map((await Promise.all(promises)).map(obj => [obj.key, obj.value]));
-
-    return [hints, treeSnapshots];
-  }
-
-  // Executes the base rollup circuit and stored the output as intermediate state for the parent merge/root circuit
-  // Executes the next level of merge if all inputs are available
-  private enqueueBaseRollup(provingState: BlockProvingState, txIndex: number) {
-    if (!provingState.verifyState()) {
-      this.logger.debug('Not running base rollup, state invalid');
-      return;
-    }
-
-    if (!provingState.tryStartProvingBase(txIndex)) {
-      this.logger.debug(`Base rollup for tx ${txIndex} already started.`);
-      return;
-    }
-
-    const txProvingState = provingState.getTxProvingState(txIndex);
-    const { processedTx } = txProvingState;
-    const { rollupType, inputs } = txProvingState.getBaseRollupTypeAndInputs();
-
-    this.logger.debug(`Enqueuing deferred proving base rollup for ${processedTx.hash.toString()}`);
-
-    this.deferredProving(
-      provingState,
-      wrapCallbackInSpan(
-        this.tracer,
-        `ProvingOrchestrator.prover.${
-          inputs instanceof PrivateTxBaseRollupPrivateInputs
-            ? 'getPrivateTxBaseRollupProof'
-            : 'getPublicTxBaseRollupProof'
-        }`,
-        {
-          [Attributes.TX_HASH]: processedTx.hash.toString(),
-          [Attributes.PROTOCOL_CIRCUIT_NAME]: rollupType,
-        },
-        signal => {
-          if (inputs instanceof PrivateTxBaseRollupPrivateInputs) {
-            return this.prover.getPrivateTxBaseRollupProof(inputs, signal, provingState.epochNumber);
-          } else {
-            return this.prover.getPublicTxBaseRollupProof(inputs, signal, provingState.epochNumber);
-          }
-        },
-      ),
-      result => {
-        this.logger.debug(`Completed proof for ${rollupType} for tx ${processedTx.hash.toString()}`);
-        validatePartialState(result.inputs.endTreeSnapshots, txProvingState.treeSnapshots);
-        const leafLocation = provingState.setBaseRollupProof(txIndex, result);
-        if (provingState.totalNumTxs === 1) {
-          this.checkAndEnqueueBlockRootRollup(provingState);
-        } else {
-          this.checkAndEnqueueNextMergeRollup(provingState, leafLocation);
-        }
-      },
-    );
-  }
-
-  // Enqueues the public chonk verifier circuit for a given transaction index, or reuses the one already enqueued.
-  // Once completed, will enqueue the the public tx base rollup.
-  protected getOrEnqueueChonkVerifier(provingState: BlockProvingState, txIndex: number) {
-    if (!provingState.verifyState()) {
-      this.logger.debug('Not running chonk verifier circuit, state invalid');
-      return;
-    }
-
-    const txProvingState = provingState.getTxProvingState(txIndex);
-    const txHash = txProvingState.processedTx.hash.toString();
-    const handleResult = (
-      result: PublicInputsAndRecursiveProof<
-        PublicChonkVerifierPublicInputs,
-        typeof NESTED_RECURSIVE_ROLLUP_HONK_PROOF_LENGTH
-      >,
-    ) => {
-      this.logger.debug(`Got chonk verifier proof for tx index: ${txIndex}`, { txHash });
-      txProvingState.setPublicChonkVerifierProof(result);
-      this.provingState?.cachedChonkVerifierProofs.delete(txHash);
-      this.checkAndEnqueueBaseRollup(provingState, txIndex);
-    };
-
-    if (this.provingState?.cachedChonkVerifierProofs.has(txHash)) {
-      this.logger.debug(`Chonk verifier proof already enqueued for tx index: ${txIndex}`, { txHash });
-      void this.provingState!.cachedChonkVerifierProofs.get(txHash)!.then(handleResult);
-      return;
-    }
-
-    this.logger.debug(`Enqueuing chonk verifier circuit for tx index: ${txIndex}`);
-    this.doEnqueueChonkVerifier(txHash, txProvingState.getPublicChonkVerifierPrivateInputs(), handleResult);
-  }
-
-  private doEnqueueChonkVerifier(
-    txHash: string,
-    inputs: PublicChonkVerifierPrivateInputs,
-    handler: (
-      result: PublicInputsAndRecursiveProof<
-        PublicChonkVerifierPublicInputs,
-        typeof NESTED_RECURSIVE_ROLLUP_HONK_PROOF_LENGTH
-      >,
-    ) => void,
-    provingState: EpochProvingState | BlockProvingState = this.provingState!,
-  ) {
-    if (!provingState.verifyState()) {
-      this.logger.debug('Not running chonk verifier circuit, state invalid');
-      return;
-    }
-
-    this.deferredProving(
-      provingState,
-      wrapCallbackInSpan(
-        this.tracer,
-        'ProvingOrchestrator.prover.getPublicChonkVerifierProof',
-        {
-          [Attributes.TX_HASH]: txHash,
-          [Attributes.PROTOCOL_CIRCUIT_NAME]: 'chonk-verifier-public' satisfies CircuitName,
-        },
-        signal => this.prover.getPublicChonkVerifierProof(inputs, signal, provingState.epochNumber),
-      ),
-      handler,
-    );
-  }
-
-  // Executes the merge rollup circuit and stored the output as intermediate state for the parent merge/block root circuit
-  // Enqueues the next level of merge if all inputs are available
-  private enqueueMergeRollup(provingState: BlockProvingState, location: TreeNodeLocation) {
-    if (!provingState.verifyState()) {
-      this.logger.debug('Not running merge rollup. State no longer valid.');
-      return;
-    }
-
-    if (!provingState.tryStartProvingMerge(location)) {
-      this.logger.debug('Merge rollup already started.');
-      return;
-    }
-
-    const inputs = provingState.getMergeRollupInputs(location);
-
-    this.deferredProving(
-      provingState,
-      wrapCallbackInSpan(
-        this.tracer,
-        'ProvingOrchestrator.prover.getTxMergeRollupProof',
-        {
-          [Attributes.PROTOCOL_CIRCUIT_NAME]: 'rollup-tx-merge' satisfies CircuitName,
-        },
-        signal => this.prover.getTxMergeRollupProof(inputs, signal, provingState.epochNumber),
-      ),
-      result => {
-        provingState.setMergeRollupProof(location, result);
-        this.checkAndEnqueueNextMergeRollup(provingState, location);
-      },
-    );
-  }
-
-  // Executes the block root rollup circuit
-  private enqueueBlockRootRollup(provingState: BlockProvingState) {
-    if (!provingState.verifyState()) {
-      this.logger.debug('Not running block root rollup, state no longer valid');
-      return;
-    }
-
-    if (!provingState.tryStartProvingBlockRoot()) {
-      this.logger.debug('Block root rollup already started.');
-      return;
-    }
-
-    const { rollupType, inputs } = provingState.getBlockRootRollupTypeAndInputs();
-
-    this.logger.debug(`Enqueuing ${rollupType} for block ${provingState.blockNumber}.`);
-
-    this.deferredProving(
-      provingState,
-      wrapCallbackInSpan(
-        this.tracer,
-        'ProvingOrchestrator.prover.getBlockRootRollupProof',
-        {
-          [Attributes.PROTOCOL_CIRCUIT_NAME]: rollupType,
-        },
-        signal => {
-          if (inputs instanceof BlockRootFirstRollupPrivateInputs) {
-            return this.prover.getBlockRootFirstRollupProof(inputs, signal, provingState.epochNumber);
-          } else if (inputs instanceof BlockRootSingleTxFirstRollupPrivateInputs) {
-            return this.prover.getBlockRootSingleTxFirstRollupProof(inputs, signal, provingState.epochNumber);
-          } else if (inputs instanceof BlockRootEmptyTxFirstRollupPrivateInputs) {
-            return this.prover.getBlockRootEmptyTxFirstRollupProof(inputs, signal, provingState.epochNumber);
-          } else if (inputs instanceof BlockRootSingleTxRollupPrivateInputs) {
-            return this.prover.getBlockRootSingleTxRollupProof(inputs, signal, provingState.epochNumber);
-          } else {
-            return this.prover.getBlockRootRollupProof(inputs, signal, provingState.epochNumber);
-          }
-        },
-      ),
-      async result => {
-        this.logger.debug(`Completed ${rollupType} proof for block ${provingState.blockNumber}`, {
-          blockNumber: provingState.blockNumber,
-          checkpointIndex: provingState.parentCheckpoint.index,
-          ...result.inputs.toInspect(),
-        });
-
-        const leafLocation = provingState.setBlockRootRollupProof(result);
-        const checkpointProvingState = provingState.parentCheckpoint;
-
-        // Verification is called from both here and setBlockCompleted. Whichever runs last
-        // will be the first to see all three pieces (header, proof output, archive) and run the checks.
-        await this.verifyBuiltBlockAgainstSyncedState(provingState);
-
-        if (checkpointProvingState.totalNumBlocks === 1) {
-          await this.checkAndEnqueueCheckpointRootRollup(checkpointProvingState);
-        } else {
-          await this.checkAndEnqueueNextBlockMergeRollup(checkpointProvingState, leafLocation);
-        }
-      },
-    );
-  }
-
-  // Executes the base parity circuit and stores the intermediate state for the root parity circuit
-  // Enqueues the root parity circuit if all inputs are available
-  private enqueueBaseParityCircuit(
-    checkpointProvingState: CheckpointProvingState,
-    provingState: BlockProvingState,
-    baseParityIndex: number,
-  ) {
-    if (!provingState.verifyState()) {
-      this.logger.debug('Not running base parity. State no longer valid.');
-      return;
-    }
-
-    if (!provingState.tryStartProvingBaseParity(baseParityIndex)) {
-      this.logger.warn(`Base parity ${baseParityIndex} already started.`);
-      return;
-    }
-
-    const inputs = checkpointProvingState.getBaseParityInputs(baseParityIndex);
-
-    this.deferredProving(
-      provingState,
-      wrapCallbackInSpan(
-        this.tracer,
-        'ProvingOrchestrator.prover.getBaseParityProof',
-        {
-          [Attributes.PROTOCOL_CIRCUIT_NAME]: 'parity-base' satisfies CircuitName,
-        },
-        signal => this.prover.getBaseParityProof(inputs, signal, provingState.epochNumber),
-      ),
-      provingOutput => {
-        provingState.setBaseParityProof(baseParityIndex, provingOutput);
-        this.checkAndEnqueueRootParityCircuit(provingState);
-      },
-    );
-  }
-
-  private checkAndEnqueueRootParityCircuit(provingState: BlockProvingState) {
-    if (!provingState.isReadyForRootParity()) {
-      return;
-    }
-
-    this.enqueueRootParityCircuit(provingState);
-  }
-
-  // Runs the root parity circuit ans stored the outputs
-  // Enqueues the root rollup proof if all inputs are available
-  private enqueueRootParityCircuit(provingState: BlockProvingState) {
-    if (!provingState.verifyState()) {
-      this.logger.debug('Not running root parity. State no longer valid.');
-      return;
-    }
-
-    if (!provingState.tryStartProvingRootParity()) {
-      this.logger.debug('Root parity already started.');
-      return;
-    }
-
-    const inputs = provingState.getParityRootInputs();
-
-    this.deferredProving(
-      provingState,
-      wrapCallbackInSpan(
-        this.tracer,
-        'ProvingOrchestrator.prover.getRootParityProof',
-        {
-          [Attributes.PROTOCOL_CIRCUIT_NAME]: 'parity-root' satisfies CircuitName,
-        },
-        signal => this.prover.getRootParityProof(inputs, signal, provingState.epochNumber),
-      ),
-      result => {
-        provingState.setRootParityProof(result);
-        this.checkAndEnqueueBlockRootRollup(provingState);
-      },
-    );
-  }
-
-  // Executes the block merge rollup circuit and stored the output as intermediate state for the parent merge/block root circuit
-  // Enqueues the next level of merge if all inputs are available
-  private enqueueBlockMergeRollup(provingState: CheckpointProvingState, location: TreeNodeLocation) {
-    if (!provingState.verifyState()) {
-      this.logger.debug('Not running block merge rollup. State no longer valid.');
-      return;
-    }
-
-    if (!provingState.tryStartProvingBlockMerge(location)) {
-      this.logger.debug('Block merge rollup already started.');
-      return;
-    }
-
-    const inputs = provingState.getBlockMergeRollupInputs(location);
-    this.deferredProving(
-      provingState,
-      wrapCallbackInSpan(
-        this.tracer,
-        'ProvingOrchestrator.prover.getBlockMergeRollupProof',
-        {
-          [Attributes.PROTOCOL_CIRCUIT_NAME]: 'rollup-block-merge' satisfies CircuitName,
-        },
-        signal => this.prover.getBlockMergeRollupProof(inputs, signal, provingState.epochNumber),
-      ),
-      async result => {
-        this.logger.debug(`Completed block merge rollup proof for checkpoint ${provingState.index}`, {
-          checkpointIndex: provingState.index,
-          mergeLocation: location,
-          ...result.inputs.toInspect(),
-        });
-        provingState.setBlockMergeRollupProof(location, result);
-        await this.checkAndEnqueueNextBlockMergeRollup(provingState, location);
-      },
-    );
-  }
-
-  private async enqueueCheckpointRootRollup(provingState: CheckpointProvingState) {
-    if (!provingState.verifyState()) {
-      this.logger.debug('Not running checkpoint root rollup. State no longer valid.');
-      return;
-    }
-
-    if (!provingState.tryStartProvingCheckpointRoot()) {
-      this.logger.debug('Checkpoint root rollup already started.');
-      return;
-    }
-
-    const rollupType = provingState.getCheckpointRootRollupType();
-
-    this.logger.debug(`Enqueuing ${rollupType} for checkpoint ${provingState.index}.`);
-
-    const inputs = await provingState.getCheckpointRootRollupInputs();
-
-    this.deferredProving(
-      provingState,
-      wrapCallbackInSpan(
-        this.tracer,
-        'ProvingOrchestrator.prover.getCheckpointRootRollupProof',
-        {
-          [Attributes.PROTOCOL_CIRCUIT_NAME]: rollupType,
-        },
-        signal => {
-          if (inputs instanceof CheckpointRootSingleBlockRollupPrivateInputs) {
-            return this.prover.getCheckpointRootSingleBlockRollupProof(inputs, signal, provingState.epochNumber);
-          } else {
-            return this.prover.getCheckpointRootRollupProof(inputs, signal, provingState.epochNumber);
-          }
-        },
-      ),
-      result => {
-        const computedEndBlobAccumulatorState = provingState.getEndBlobAccumulator()!.toBlobAccumulator();
-        const circuitEndBlobAccumulatorState = result.inputs.endBlobAccumulator;
-        if (!circuitEndBlobAccumulatorState.equals(computedEndBlobAccumulatorState)) {
-          this.logger.error(
-            `Blob accumulator state mismatch.\nCircuit: ${inspect(circuitEndBlobAccumulatorState)}\nComputed: ${inspect(
-              computedEndBlobAccumulatorState,
-            )}`,
-          );
-          provingState.reject(`Blob accumulator state mismatch.`);
-          return;
-        }
-
-        this.logger.debug(`Completed ${rollupType} proof for checkpoint ${provingState.index}`, {
-          checkpointIndex: provingState.index,
-          ...result.inputs.toInspect(),
-        });
-
-        const leafLocation = provingState.setCheckpointRootRollupProof(result);
-        const epochProvingState = provingState.parentEpoch;
-
-        if (epochProvingState.totalNumCheckpoints === 1) {
-          this.enqueueEpochPadding(epochProvingState);
-        } else {
-          this.checkAndEnqueueNextCheckpointMergeRollup(epochProvingState, leafLocation);
-        }
-      },
-    );
-  }
-
-  private checkAndEnqueueNextMergeRollup(provingState: BlockProvingState, currentLocation: TreeNodeLocation) {
-    if (!provingState.isReadyForMergeRollup(currentLocation)) {
-      return;
-    }
-
-    const parentLocation = provingState.getParentLocation(currentLocation);
-    if (parentLocation.level === 0) {
-      this.checkAndEnqueueBlockRootRollup(provingState);
-    } else {
-      this.enqueueMergeRollup(provingState, parentLocation);
-    }
-  }
-
-  private checkAndEnqueueBlockRootRollup(provingState: BlockProvingState) {
-    if (!provingState.isReadyForBlockRootRollup()) {
-      this.logger.debug('Not ready for block root rollup');
-      return;
-    }
-
-    this.enqueueBlockRootRollup(provingState);
-  }
-
-  private async checkAndEnqueueNextBlockMergeRollup(
-    provingState: CheckpointProvingState,
-    currentLocation: TreeNodeLocation,
-  ) {
-    if (!provingState.isReadyForBlockMerge(currentLocation)) {
-      return;
-    }
-
-    const parentLocation = provingState.getParentLocation(currentLocation);
-    if (parentLocation.level === 0) {
-      await this.checkAndEnqueueCheckpointRootRollup(provingState);
-    } else {
-      this.enqueueBlockMergeRollup(provingState, parentLocation);
-    }
-  }
-
-  protected async checkAndEnqueueCheckpointRootRollup(provingState: CheckpointProvingState) {
-    if (!provingState.isReadyForCheckpointRoot()) {
-      return;
-    }
-
-    await this.enqueueCheckpointRootRollup(provingState);
-  }
-
-  /**
-   * Executes the VM circuit for a public function, will enqueue the corresponding kernel if the
-   * previous kernel is ready
-   * @param provingState - The proving state being operated on
-   * @param txIndex - The index of the transaction being proven
-   */
-  private enqueueVM(provingState: BlockProvingState, txIndex: number) {
-    if (!provingState.verifyState()) {
-      this.logger.debug(`Not running VM circuit as state is no longer valid`);
-      return;
-    }
-
-    const txProvingState = provingState.getTxProvingState(txIndex);
-
-    const doAvmProving = wrapCallbackInSpan(
-      this.tracer,
-      'ProvingOrchestrator.prover.getAvmProof',
-      {
-        [Attributes.TX_HASH]: txProvingState.processedTx.hash.toString(),
-      },
-      async (signal: AbortSignal) => {
-        const inputs = txProvingState.getAvmInputs();
-        return await this.prover.getAvmProof(inputs, signal, provingState.epochNumber);
-      },
-    );
-
-    this.deferredProving(provingState, doAvmProving, proof => {
-      this.logger.debug(`Proven VM for tx index: ${txIndex}`);
-      txProvingState.setAvmProof(proof);
-      this.checkAndEnqueueBaseRollup(provingState, txIndex);
-    });
-  }
-
-  protected checkAndEnqueueBaseRollup(provingState: BlockProvingState, txIndex: number) {
-    const txProvingState = provingState.getTxProvingState(txIndex);
-    if (!txProvingState.ready()) {
-      return;
-    }
-
-    // We must have completed all proving (chonk verifier proof and (if required) vm proof are generated), we now move to the base rollup.
-    this.logger.debug(`Public functions completed for tx ${txIndex} enqueueing base rollup`);
-
-    this.enqueueBaseRollup(provingState, txIndex);
-  }
-}
