--- yarn-project/prover-client/src/orchestrator/orchestrator_workflow.test.ts
@@ -1,318 +1,0 @@
-import { NESTED_RECURSIVE_PROOF_LENGTH, RECURSIVE_PROOF_LENGTH } from '@aztec/constants';
-import { EpochNumber } from '@aztec/foundation/branded-types';
-import { timesAsync } from '@aztec/foundation/collection';
-import { createLogger } from '@aztec/foundation/log';
-import { promiseWithResolvers } from '@aztec/foundation/promise';
-import { retryFastUntil } from '@aztec/foundation/retry';
-import { ProtocolCircuitVks } from '@aztec/noir-protocol-circuits-types/server/vks';
-import {
-  type PublicInputsAndRecursiveProof,
-  type ServerCircuitProver,
-  makePublicInputsAndRecursiveProof,
-} from '@aztec/stdlib/interfaces/server';
-import type { ParityPublicInputs } from '@aztec/stdlib/parity';
-import { ChonkProof, makeRecursiveProof } from '@aztec/stdlib/proofs';
-import { makeParityPublicInputs } from '@aztec/stdlib/testing';
-import { Tx } from '@aztec/stdlib/tx';
-
-import { jest } from '@jest/globals';
-import { type MockProxy, mock } from 'jest-mock-extended';
-
-import { TestContext } from '../mocks/test_context.js';
-import type { ProvingOrchestrator } from './orchestrator.js';
-
-const logger = createLogger('prover-client:test:orchestrator-workflow');
-
-describe('prover/orchestrator', () => {
-  describe('workflow', () => {
-    let orchestrator: ProvingOrchestrator;
-    let context: TestContext;
-
-    afterEach(async () => {
-      await context.cleanup();
-    });
-
-    describe('with mock prover', () => {
-      let mockProver: MockProxy<ServerCircuitProver>;
-
-      beforeEach(async () => {
-        mockProver = mock<ServerCircuitProver>();
-        context = await TestContext.new(logger, {
-          proverCount: 4,
-          createProver: () => Promise.resolve(mockProver),
-        });
-        ({ orchestrator } = context);
-      });
-
-      it('calls root parity circuit only when ready', async () => {
-        // create a custom L2 to L1 message
-        const numL1ToL2Messages = 1;
-        const {
-          constants,
-          blocks: [{ header }],
-          l1ToL2Messages,
-          previousBlockHeader,
-        } = await context.makeCheckpoint(1, { numTxsPerBlock: 0, numL1ToL2Messages });
-
-        const message = l1ToL2Messages[0];
-
-        // and delay its proof
-        const pendingBaseParityResult =
-          promiseWithResolvers<PublicInputsAndRecursiveProof<ParityPublicInputs, typeof RECURSIVE_PROOF_LENGTH>>();
-        const expectedBaseParityResult = makePublicInputsAndRecursiveProof(
-          makeParityPublicInputs(0xff),
-          makeRecursiveProof(RECURSIVE_PROOF_LENGTH),
-          ProtocolCircuitVks.ParityBaseArtifact,
-        );
-
-        mockProver.getRootParityProof.mockResolvedValue(
-          makePublicInputsAndRecursiveProof(
-            makeParityPublicInputs(),
-            makeRecursiveProof(NESTED_RECURSIVE_PROOF_LENGTH),
-            ProtocolCircuitVks.ParityRootArtifact,
-          ),
-        );
-
-        mockProver.getBaseParityProof.mockImplementation(inputs => {
-          if (inputs.msgs[0].equals(message)) {
-            return pendingBaseParityResult.promise;
-          } else {
-            return Promise.resolve(
-              makePublicInputsAndRecursiveProof(
-                makeParityPublicInputs(),
-                makeRecursiveProof(RECURSIVE_PROOF_LENGTH),
-                ProtocolCircuitVks.ParityBaseArtifact,
-              ),
-            );
-          }
-        });
-
-        const finalBlobChallenges = await context.getFinalBlobChallenges();
-        orchestrator.startNewEpoch(EpochNumber(1), 1, finalBlobChallenges);
-
-        await orchestrator.startNewCheckpoint(
-          0, // checkpointIndex
-          constants,
-          [message],
-          1,
-          previousBlockHeader,
-        );
-
-        const { blockNumber, timestamp } = header.globalVariables;
-        await orchestrator.startNewBlock(blockNumber, timestamp, 0 /* numTxs */);
-
-        // the prover broker deduplicates jobs, so the base parity proof
-        // for the three sets empty messages is called only once. so total
-        // calls is one for the empty messages and one for the custom message.
-        await retryFastUntil(
-          () => mockProver.getBaseParityProof.mock.calls.length === 2,
-          'base parity proofs to be called',
-        );
-        expect(mockProver.getRootParityProof).not.toHaveBeenCalled();
-
-        // only after the base parity proof is resolved, the root parity should be called
-        pendingBaseParityResult.resolve(expectedBaseParityResult);
-
-        // give the orchestrator a chance to calls its callbacks
-        await retryFastUntil(
-          () => mockProver.getRootParityProof.mock.calls.length === 1,
-          'root parity proof to be called',
-        );
-
-        orchestrator.cancel();
-      });
-    });
-
-    describe('with simulated prover', () => {
-      let prover: ServerCircuitProver;
-
-      beforeEach(async () => {
-        context = await TestContext.new(logger);
-        ({ prover, orchestrator } = context);
-      });
-
-      it('waits for block to be completed before enqueueing block root proof', async () => {
-        const numBlocks = 1;
-        const {
-          constants,
-          blocks: [{ header, txs }],
-          previousBlockHeader,
-        } = await context.makeCheckpoint(numBlocks);
-
-        const finalBlobChallenges = await context.getFinalBlobChallenges();
-        orchestrator.startNewEpoch(EpochNumber(1), 1, finalBlobChallenges);
-
-        await orchestrator.startNewCheckpoint(
-          0, // checkpointIndex
-          constants,
-          [],
-          numBlocks,
-          previousBlockHeader,
-        );
-
-        const { blockNumber, timestamp } = header.globalVariables;
-        await orchestrator.startNewBlock(blockNumber, timestamp, txs.length);
-
-        // now finish the block
-        await orchestrator.setBlockCompleted(blockNumber);
-
-        const result = await orchestrator.finalizeEpoch();
-        expect(result.proof).toBeDefined();
-      });
-
-      it('cleans up all world state forks', async () => {
-        const numBlocks = 1;
-        const {
-          constants,
-          blocks: [{ header, txs }],
-          previousBlockHeader,
-        } = await context.makeCheckpoint(numBlocks);
-
-        const finalBlobChallenges = await context.getFinalBlobChallenges();
-        orchestrator.startNewEpoch(EpochNumber(1), 1, finalBlobChallenges);
-
-        await orchestrator.startNewCheckpoint(
-          0, // checkpointIndex
-          constants,
-          [],
-          numBlocks,
-          previousBlockHeader,
-        );
-
-        const { blockNumber, timestamp } = header.globalVariables;
-        await orchestrator.startNewBlock(blockNumber, timestamp, txs.length);
-
-        // now finish the block
-        await orchestrator.setBlockCompleted(blockNumber);
-
-        const result = await orchestrator.finalizeEpoch();
-        expect(result.proof).toBeDefined();
-        // Forks are closed deterministically in setBlockCompleted, so no cancel() needed.
-        expect(orchestrator.getNumActiveForks()).toEqual(0);
-      });
-
-      it('can start chonk verifier proofs before adding processed txs', async () => {
-        const getChonkVerifierSpy = jest.spyOn(prover, 'getPublicChonkVerifierProof');
-
-        const numBlocks = 1;
-        const {
-          constants,
-          blocks: [{ header, txs }],
-          previousBlockHeader,
-        } = await context.makeCheckpoint(numBlocks, {
-          numTxsPerBlock: 2,
-          makeProcessedTxOpts: () => ({ privateOnly: false }), // The chonk verifier circuit is only used for public txs.
-        });
-
-        const finalBlobChallenges = await context.getFinalBlobChallenges();
-        orchestrator.startNewEpoch(EpochNumber(1), 1, finalBlobChallenges);
-
-        await orchestrator.startNewCheckpoint(
-          0, // checkpointIndex
-          constants,
-          [],
-          numBlocks,
-          previousBlockHeader,
-        );
-
-        txs.forEach(tx => (tx.chonkProof = ChonkProof.random()));
-        await orchestrator.startChonkVerifierCircuits(
-          txs.map(tx =>
-            Tx.from({
-              txHash: tx.hash,
-              data: tx.data,
-              chonkProof: tx.chonkProof,
-              contractClassLogFields: [],
-              publicFunctionCalldata: [],
-            }),
-          ),
-        );
-
-        await retryFastUntil(() => getChonkVerifierSpy.mock.calls.length === 2, 'chonk verifier proofs to be called');
-        getChonkVerifierSpy.mockReset();
-
-        const { blockNumber, timestamp } = header.globalVariables;
-        await orchestrator.startNewBlock(blockNumber, timestamp, txs.length);
-        await orchestrator.addTxs(txs);
-        await orchestrator.setBlockCompleted(blockNumber);
-
-        const result = await orchestrator.finalizeEpoch();
-        expect(result.proof).toBeDefined();
-        expect(getChonkVerifierSpy).toHaveBeenCalledTimes(0);
-      });
-
-      it('can add checkpoints in arbitrary order', async () => {
-        const numCheckpoints = 3;
-        const numBlocksPerCheckpoint = 2;
-        const numTxsPerBlock = 2;
-        const checkpoints = await timesAsync(numCheckpoints, () =>
-          context.makeCheckpoint(numBlocksPerCheckpoint, {
-            numTxsPerBlock,
-          }),
-        );
-
-        const finalBlobChallenges = await context.getFinalBlobChallenges();
-        context.orchestrator.startNewEpoch(EpochNumber(1), numCheckpoints, finalBlobChallenges);
-
-        // Start checkpoint in reverse order.
-        for (let checkpointIndex = numCheckpoints - 1; checkpointIndex >= 0; checkpointIndex--) {
-          const { constants, blocks, l1ToL2Messages, previousBlockHeader } = checkpoints[checkpointIndex];
-          await context.orchestrator.startNewCheckpoint(
-            checkpointIndex,
-            constants,
-            l1ToL2Messages,
-            blocks.length,
-            previousBlockHeader,
-          );
-
-          // Blocks in a checkpoint need to be started in order.
-          for (const block of blocks) {
-            const { txs } = block;
-            const { blockNumber, timestamp } = block.header.globalVariables;
-            await context.orchestrator.startNewBlock(blockNumber, timestamp, txs.length);
-            await context.orchestrator.addTxs(txs);
-            await context.orchestrator.setBlockCompleted(blockNumber);
-          }
-        }
-
-        logger.info('Finalizing epoch');
-        const epoch = await context.orchestrator.finalizeEpoch();
-        expect(epoch.proof).toBeDefined();
-      });
-
-      it('can add checkpoints asynchronously', async () => {
-        const numCheckpoints = 4;
-        const numBlocksPerCheckpoint = 2;
-        const numTxsPerBlock = 1;
-        const checkpoints = await timesAsync(numCheckpoints, () =>
-          context.makeCheckpoint(numBlocksPerCheckpoint, { numTxsPerBlock }),
-        );
-
-        const finalBlobChallenges = await context.getFinalBlobChallenges();
-        context.orchestrator.startNewEpoch(EpochNumber(1), numCheckpoints, finalBlobChallenges);
-
-        await Promise.all(
-          checkpoints.map(async (checkpoint, checkpointIndex) => {
-            const { constants, blocks, l1ToL2Messages, previousBlockHeader } = checkpoint;
-            await context.orchestrator.startNewCheckpoint(
-              checkpointIndex,
-              constants,
-              l1ToL2Messages,
-              blocks.length,
-              previousBlockHeader,
-            );
-
-            // Blocks in a checkpoint need to be added in order.
-            for (const block of blocks) {
-              const { txs } = block;
-              const { blockNumber, timestamp } = block.header.globalVariables;
-              await context.orchestrator.startNewBlock(blockNumber, timestamp, txs.length);
-              await context.orchestrator.addTxs(txs);
-              await context.orchestrator.setBlockCompleted(blockNumber);
-            }
-          }),
-        );
-      });
-    });
-  });
-});
