+++ yarn-project/prover-client/src/orchestrator/proving-scheduler.ts
@@ -4,9 +4,9 @@ import { SerialQueue } from '@aztec/foundation/queue';
 import { sleep } from '@aztec/foundation/sleep';
 
 /**
- * Minimal surface a deferred-proving state must expose. Both `EpochProvingState` /
- * `CheckpointProvingState` / `BlockProvingState` (used by `ProvingOrchestrator`) and
- * `TopTreeProvingState` (used by `TopTreeOrchestrator`) satisfy it.
+ * Minimal surface a deferred-proving state must expose. Both `CheckpointProvingState` /
+ * `BlockProvingState` (used by `CheckpointSubTreeOrchestrator`) and `TopTreeProvingState`
+ * (used by `TopTreeOrchestrator`) satisfy it.
  */
 export interface ProvingStateLike {
   /** Returns false once the state has been cancelled or otherwise invalidated. */
@@ -19,7 +19,10 @@ export interface ProvingStateLike {
  * Common scheduling infrastructure shared by every orchestrator that drives broker
  * proving jobs:
  *
- *   - One `SerialQueue` (`deferredJobQueue`) acting as the enqueue-throttle.
+ *   - A shared `SerialQueue` (`deferredJobQueue`) acting as the enqueue-throttle. The
+ *     queue is owned by the `ProverClient` and shared across every orchestrator (every
+ *     sub-tree and top-tree across every concurrent epoch session), so the total rate
+ *     of job submission to the broker is bounded once, not once-per-orchestrator.
  *   - A list of `AbortController`s (`pendingProvingJobs`) so a `cancel()` can abort
  *     in-flight broker jobs when needed.
  *   - A `deferredProving<T>(state, request, callback, isCancelled?)` method that wraps
@@ -28,23 +31,20 @@ export interface ProvingStateLike {
  *
  * Subclasses own their own concrete proving state and define `cancelInternal()` for
  * the rest of the cleanup work (closing world-state forks, marking sub-trees
- * cancelled, etc.). `stop()` lives on the base class and follows the standard pattern
- * of grabbing the old queue, calling `cancelInternal()` (which recreates the queue),
- * and awaiting the old queue's drain.
+ * cancelled, etc.). Because the queue is shared, neither `cancel()` nor `stop()` touch
+ * it — they only abort this orchestrator's in-flight broker jobs. Queued-but-unrun jobs
+ * for a cancelled orchestrator no-op via the guards in `deferredProving`.
  */
 export abstract class ProvingScheduler {
   protected pendingProvingJobs: AbortController[] = [];
   protected logger: Logger;
-  private deferredJobQueue: SerialQueue;
 
   constructor(
-    private readonly enqueueConcurrency: number,
+    private readonly deferredJobQueue: SerialQueue,
     loggerName = 'prover-client:proving-scheduler',
     bindings?: LoggerBindings,
   ) {
     this.logger = createLogger(loggerName, bindings);
-    this.deferredJobQueue = new SerialQueue();
-    this.deferredJobQueue.start(this.enqueueConcurrency);
   }
 
   /** Number of broker jobs currently in flight. */
@@ -53,16 +53,14 @@ export abstract class ProvingScheduler {
   }
 
   /**
-   * Drains the deferred-job queue, recreates it (so the subclass can be reused), and
-   * optionally aborts every in-flight broker job. Aborting is the right choice on
+   * Optionally aborts every in-flight broker job. Aborting is the right choice on
    * reorg-driven cancel (where the in-flight inputs are no longer valid) and the
    * wrong choice on shutdown (where leaving jobs in the broker queue lets a restart
-   * pick them up).
+   * pick them up). The shared queue is not touched — queued-but-unrun jobs belonging
+   * to this orchestrator no-op once their controller is aborted (or once their state
+   * is marked invalid by the subclass).
    */
   protected resetSchedulerState(abortJobs: boolean): void {
-    void this.deferredJobQueue.cancel();
-    this.deferredJobQueue = new SerialQueue();
-    this.deferredJobQueue.start(this.enqueueConcurrency);
     if (abortJobs) {
       for (const controller of this.pendingProvingJobs) {
         controller.abort();
@@ -78,14 +76,12 @@ export abstract class ProvingScheduler {
   protected abstract cancelInternal(): void;
 
   /**
-   * Standard stop: grab the old queue, cancel (which recreates the queue), then
-   * await the old queue's drain so any final job tear-down has unwound before we
-   * return.
+   * Standard stop: cancel this orchestrator's work. The shared queue is owned by the
+   * `ProverClient` and outlives every orchestrator, so it is not drained here.
    */
-  public async stop(): Promise<void> {
-    const oldQueue = this.deferredJobQueue;
+  public stop(): Promise<void> {
     this.cancelInternal();
-    await oldQueue.cancel();
+    return Promise.resolve();
   }
 
   /**
