+++ yarn-project/prover-client/src/orchestrator/top-tree-orchestrator.test.ts
@@ -1,14 +1,28 @@
+import { MAX_L2_TO_L1_MSGS_PER_TX } from '@aztec/constants';
 import { EpochNumber } from '@aztec/foundation/branded-types';
+import { padArrayEnd } from '@aztec/foundation/collection';
 import { EthAddress } from '@aztec/foundation/eth-address';
 import { createLogger } from '@aztec/foundation/log';
 import { promiseWithResolvers } from '@aztec/foundation/promise';
+import { retryUntil } from '@aztec/foundation/retry';
+import { sleep } from '@aztec/foundation/sleep';
+import { ScopedL2ToL1Message, computeEpochOutHash } from '@aztec/stdlib/messaging';
+import { makeScopedL2ToL1Message } from '@aztec/stdlib/testing';
 
-import { TestContext } from '../mocks/test_context.js';
+import { TestContext, makeTestDeferredJobQueue } from '../mocks/test_context.js';
 import { CheckpointSubTreeOrchestrator } from './checkpoint-sub-tree-orchestrator.js';
-import { EpochProvingContext } from './epoch-proving-context.js';
+import { ChonkCache } from './chonk-cache.js';
 import { type CheckpointTopTreeData, TopTreeCancelledError, TopTreeOrchestrator } from './top-tree-orchestrator.js';
 
 const logger = createLogger('prover-client:test:top-tree-orchestrator');
+
+/** A full tx-worth of L2-to-L1 messages, padded to the per-tx maximum. */
+const makeL2ToL1Messages = (count: number) =>
+  padArrayEnd(
+    Array.from({ length: count }, (_, i) => makeScopedL2ToL1Message((i + 1) * 321)),
+    ScopedL2ToL1Message.empty(),
+    MAX_L2_TO_L1_MSGS_PER_TX,
+  );
 
 /**
  * End-to-end exercises for `TopTreeOrchestrator`. Each test drives one or more
@@ -31,17 +45,24 @@ describe('prover/orchestrator/top-tree', () => {
    * Drives a single checkpoint through `CheckpointSubTreeOrchestrator` and returns
    * the assembled `CheckpointTopTreeData` plus the originating checkpoint metadata.
    */
-  async function driveSubTree(numBlocks: number, numTxsPerBlock: number, numL1ToL2Messages = 0) {
-    const fixture = await context.makeCheckpoint(numBlocks, { numTxsPerBlock, numL1ToL2Messages });
+  async function driveSubTree(numBlocks: number, numTxsPerBlock: number, numL1ToL2Messages = 0, numL2ToL1Messages = 0) {
+    const fixture = await context.makeCheckpoint(numBlocks, {
+      numTxsPerBlock,
+      numL1ToL2Messages,
+      makeProcessedTxOpts:
+        numL2ToL1Messages > 0
+          ? () => ({ privateOnly: false, avmAccumulatedData: { l2ToL1Msgs: makeL2ToL1Messages(numL2ToL1Messages) } })
+          : undefined,
+    });
 
-    const epochContext = new EpochProvingContext(context.prover, EpochNumber(1));
     const subTree = await CheckpointSubTreeOrchestrator.start(
       context.worldState,
       context.prover,
       EthAddress.ZERO,
-      epochContext,
+      new ChonkCache(),
+      EpochNumber(1),
       false,
-      10,
+      makeTestDeferredJobQueue(),
       fixture.constants,
       fixture.l1ToL2Messages,
       numBlocks,
@@ -60,7 +81,6 @@ describe('prover/orchestrator/top-tree', () => {
 
     const result = await resultPromise;
     await subTree.stop();
-    epochContext.stop();
 
     const topTreeData: CheckpointTopTreeData = {
       blockProofs: Promise.resolve(result.blockProofOutputs),
@@ -77,7 +97,7 @@ describe('prover/orchestrator/top-tree', () => {
     const { topTreeData } = await driveSubTree(1, 1);
     const challenges = await context.getFinalBlobChallenges();
 
-    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, 10);
+    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, makeTestDeferredJobQueue());
     try {
       const result = await topTree.prove(EpochNumber(1), 1, challenges, [topTreeData]);
       expect(result.proof).toBeDefined();
@@ -93,10 +113,47 @@ describe('prover/orchestrator/top-tree', () => {
     const b = await driveSubTree(1, 1);
     const challenges = await context.getFinalBlobChallenges();
 
-    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, 10);
+    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, makeTestDeferredJobQueue());
     try {
       const result = await topTree.prove(EpochNumber(1), 2, challenges, [a.topTreeData, b.topTreeData]);
       expect(result.proof).toBeDefined();
+    } finally {
+      await topTree.stop();
+    }
+  });
+
+  it('produces an epoch proof for a checkpoint carrying L1-to-L2 messages', async () => {
+    // L1-to-L2 (cross-chain) messages must survive the full sub-tree → top-tree path (A-1039).
+    const { topTreeData } = await driveSubTree(1, 1, 3);
+    const challenges = await context.getFinalBlobChallenges();
+
+    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, makeTestDeferredJobQueue());
+    try {
+      const result = await topTree.prove(EpochNumber(1), 1, challenges, [topTreeData]);
+      expect(result.proof).toBeDefined();
+      expect(result.publicInputs).toBeDefined();
+    } finally {
+      await topTree.stop();
+    }
+  });
+
+  it('produces an epoch proof for a checkpoint emitting L2-to-L1 messages', async () => {
+    // L2-to-L1 messages feed the epoch out-hash assembled at the top tree (A-1039).
+    const { fixture, topTreeData } = await driveSubTree(1, 1, 0, 2);
+    expect(fixture.blocks[0].txs[0].txEffect.l2ToL1Msgs.length).toBe(2);
+    const challenges = await context.getFinalBlobChallenges();
+
+    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, makeTestDeferredJobQueue());
+    try {
+      const result = await topTree.prove(EpochNumber(1), 1, challenges, [topTreeData]);
+      expect(result.proof).toBeDefined();
+      expect(result.publicInputs).toBeDefined();
+
+      // The messages flow all the way through to the epoch out-hash on the root-rollup proof.
+      const messagesPerEpoch = [fixture.blocks.map(b => b.txs.map(tx => tx.txEffect.l2ToL1Msgs))];
+      const expectedEpochOutHash = computeEpochOutHash(messagesPerEpoch);
+      expect(expectedEpochOutHash.isZero()).toBe(false); // sanity: the fixture really did carry messages
+      expect(result.publicInputs.outHash).toEqual(expectedEpochOutHash);
     } finally {
       await topTree.stop();
     }
@@ -112,7 +169,7 @@ describe('prover/orchestrator/top-tree', () => {
     const deferred = promiseWithResolvers<typeof b.topTreeData.blockProofs extends Promise<infer T> ? T : never>();
     const ckpt1 = { ...b.topTreeData, blockProofs: deferred.promise } as CheckpointTopTreeData;
 
-    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, 10);
+    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, makeTestDeferredJobQueue());
     try {
       // Top tree proves in the background; it should be able to advance ckpt0's root
       // rollup before we resolve ckpt1's promise.
@@ -139,7 +196,7 @@ describe('prover/orchestrator/top-tree', () => {
     const stuck = new Promise<typeof topTreeData.blockProofs extends Promise<infer T> ? T : never>(() => {});
     const stuckData = { ...topTreeData, blockProofs: stuck } as CheckpointTopTreeData;
 
-    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, 10);
+    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, makeTestDeferredJobQueue());
     const provePromise = topTree.prove(EpochNumber(1), 1, challenges, [stuckData]);
 
     // Yield then cancel.
@@ -159,7 +216,7 @@ describe('prover/orchestrator/top-tree', () => {
     const { topTreeData } = await driveSubTree(1, 1);
     const challenges = await context.getFinalBlobChallenges();
 
-    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, 10);
+    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, makeTestDeferredJobQueue());
     topTree.cancel({ abortJobs: true });
 
     let actual: unknown;
@@ -176,7 +233,7 @@ describe('prover/orchestrator/top-tree', () => {
     const { topTreeData } = await driveSubTree(1, 1);
     const challenges = await context.getFinalBlobChallenges();
 
-    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, 10);
+    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, makeTestDeferredJobQueue());
     try {
       const first = topTree.prove(EpochNumber(1), 1, challenges, [topTreeData]);
       // Second call before first settles should throw synchronously inside the function
@@ -187,11 +244,86 @@ describe('prover/orchestrator/top-tree', () => {
     }
   });
 
+  it('rejects (does not hang) when building checkpoint-root inputs fails', async () => {
+    // A-1036: if input-building throws (bad block proof, blob-hint failure, etc.) the failure
+    // must reach state.reject(). Otherwise the completion promise never settles and prove() hangs.
+    const { topTreeData } = await driveSubTree(1, 1);
+    const challenges = await context.getFinalBlobChallenges();
+
+    // A malformed block proof makes toProofData (inside buildCheckpointRootInputs) throw.
+    const badData = { ...topTreeData, blockProofs: Promise.resolve([{} as any]) } as CheckpointTopTreeData;
+
+    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, makeTestDeferredJobQueue());
+    try {
+      const provePromise = topTree.prove(EpochNumber(1), 1, challenges, [badData]);
+      const hung = Symbol('hung');
+      const outcome = await Promise.race([
+        provePromise.then(
+          () => 'resolved' as const,
+          err => err,
+        ),
+        sleep(5000).then(() => hung),
+      ]);
+      expect(outcome).not.toBe(hung);
+      expect(outcome).toBeInstanceOf(Error);
+      expect((outcome as Error).message).toMatch(/checkpoint root inputs/i);
+    } finally {
+      topTree.cancel({ abortJobs: true });
+      await topTree.stop();
+    }
+  });
+
+  it('surfaces a genuine proving failure even when a cancel races in', async () => {
+    // A-1035: a real failure rejects the completion promise first, then a reorg cancel arrives
+    // before prove()'s catch observes it. The genuine error must survive, not be masked as
+    // TopTreeCancelledError.
+    const { topTreeData } = await driveSubTree(1, 1);
+    const challenges = await context.getFinalBlobChallenges();
+
+    const deferred = promiseWithResolvers<typeof topTreeData.blockProofs extends Promise<infer T> ? T : never>();
+    // Observe exactly when prove() attaches its blockProofs handler, so we can sequence the
+    // genuine rejection and the cancel deterministically rather than racing a fixed timeout.
+    let handlerAttached = false;
+    const observableBlockProofs = {
+      then: (onF: any, onR: any) => {
+        handlerAttached = true;
+        return deferred.promise.then(onF, onR);
+      },
+    };
+    const failingData = { ...topTreeData, blockProofs: observableBlockProofs as any } as CheckpointTopTreeData;
+
+    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, makeTestDeferredJobQueue());
+    const provePromise = topTree.prove(EpochNumber(1), 1, challenges, [failingData]);
+
+    // Wait until prove() has finished its pre-loop setup and registered the blockProofs handler.
+    await retryUntil(() => handlerAttached, 'prove() attaches blockProofs handler', 5, 0.005);
+
+    // Register a cancel reaction on the rejection, after prove()'s own handler (registered
+    // first, so it runs first). On rejection the ordering is: prove's handler rejects the
+    // completion promise with the genuine error → our reaction cancels → prove's catch runs.
+    // That places the cancel in the exact one-microtask window where A-1035's masking occurs.
+    const cancelOnReject = deferred.promise.catch(() => topTree.cancel({ abortJobs: true }));
+
+    // Genuine failure rejects the completion promise while cancelled is still false.
+    deferred.reject(new Error('REAL CIRCUIT FAILURE'));
+    await cancelOnReject;
+
+    const err = await provePromise.then(
+      () => undefined,
+      e => e,
+    );
+    expect(err).toBeInstanceOf(Error);
+    expect(err).not.toBeInstanceOf(TopTreeCancelledError);
+    expect((err as Error).message).toContain('REAL CIRCUIT FAILURE');
+
+    await topTree.stop();
+  });
+
   it('rejects when checkpointData length disagrees with totalNumCheckpoints', async () => {
     const { topTreeData } = await driveSubTree(1, 1);
     const challenges = await context.getFinalBlobChallenges();
 
-    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, 10);
+    const topTree = new TopTreeOrchestrator(context.prover, EthAddress.ZERO, makeTestDeferredJobQueue());
     try {
       await expect(topTree.prove(EpochNumber(1), 2, challenges, [topTreeData])).rejects.toThrow(
         /does not match totalNumCheckpoints/,
